README
This app let you to set alarm and notification text at any time you want, and it will notify you with the beutiful alarm , and moreover it will notify you even if your phone was off, and after turning it on, you will receive your notification. Thats it. Thank you:)